% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandwidth_selection.R
\name{bandwidth_selection}
\alias{bandwidth_selection}
\title{Bandwidth selection using optimise.}
\usage{
bandwidth_selection(
  criterion_getter,
  Kernel,
  data,
  lower = 1e-10,
  upper = 10,
  ...
)
}
\arguments{
\item{criterion_getter}{A function. The bandwidth selection method. Must return a function accepting exactly one numeric parameter h (the bandwidth) and returning a number. This is the optimisation criterion, smaller values are considered better.}

\item{Kernel}{A real function. The kernel.}

\item{data}{A double vector of the sample data to use.}

\item{lower}{A double vector of length 1. The lowest bandwidth to test.}

\item{upper}{A double vector of length 1. The greatest bandwidth to test.}

\item{...}{Optional arguments for the criterion function.}
}
\value{
A double vector of length 1. The optimal bandwidth.
}
\description{
Find the optimal value for the bandwidth h.
}
