% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandwidth_selection.R
\name{bandwidth_selection_criteria}
\alias{bandwidth_selection_criteria}
\title{Getters for the bandwidth selection criteria.}
\usage{
bandwidth_selection_criteria()
}
\description{
Getters for the bandwidth selection criteria.
}
