% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandwidth_selection_GL.R
\name{est_bias_GL}
\alias{est_bias_GL}
\title{Estimator for the Bias Term.}
\usage{
est_bias_GL(h, Kernel, data, maxEval, bandwidths, c)
}
\arguments{
\item{h}{A double vector of length 1. The bandwidth.}

\item{Kernel}{A real function. The kernel.}

\item{data}{A double vector of the sample data to use.}

\item{maxEval}{A double vector of length 1. The maximum number of function evaluations when integrating.}

\item{bandwidths}{A double vector. The set of h_prime bandwidths to test.}

\item{c}{A double vector of length 1. A calibration constant.}

\item{v}{A double vector of length 1. A calibration constant.}
}
\value{
A double vector of length 1.
}
\description{
B_hat
}
