% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandwidth_selection_GL.R,
%   R/bandwidth_selection_PCO.R
\name{est_risk}
\alias{est_risk}
\title{Estimator for the Risk.}
\usage{
est_risk(h, Kernel, data, m, v)

est_risk(h, Kernel, data, m, v)
}
\arguments{
\item{h}{A double vector of length 1. The bandwidth.}

\item{Kernel}{A real function. The kernel.}

\item{data}{A double vector of the sample data to use.}

\item{m}{A double vector of length 1. The smallest bandwidth.}

\item{v}{A double vector of length 1. A calibration constant for weighing the variance term.}

\item{bandwidths}{A double vector. The set of h_prime bandwidths to test.}
}
\value{
A double vector of length 1.

A double vector of length 1.
}
\description{
Risk_hat

Risk_hat
}
