% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandwidth_selection_GL.R
\name{est_risk_GL}
\alias{est_risk_GL}
\title{Estimator for the Risk.}
\usage{
est_risk_GL(h, Kernel, data, maxEval, bandwidths, c, v)
}
\arguments{
\item{h}{A double vector of length 1. The bandwidth.}

\item{Kernel}{A real function. The kernel.}

\item{data}{A double vector of the sample data to use.}

\item{maxEval}{A double vector of length 1. The maximum number of function evaluations when integrating.}

\item{bandwidths}{A double vector. The set of h_prime bandwidths to test.}

\item{c}{A double vector of length 1. A calibration constant for weighing the variance term inside the bias term.}

\item{v}{A double vector of length 1. A calibration constant for weighing the variance term.}
}
\value{
A double vector of length 1.
}
\description{
Risk_hat
}
