% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandwidth_selection_PCO.R
\name{est_risk_PCO}
\alias{est_risk_PCO}
\title{Estimator for the Risk.}
\usage{
est_risk_PCO(h, Kernel, data, maxEval, m, v)
}
\arguments{
\item{h}{A double vector of length 1. The bandwidth.}

\item{Kernel}{A real function. The kernel.}

\item{data}{A double vector of the sample data to use.}

\item{maxEval}{A double vector of length 1. The maximum number of function evaluations when integrating.}

\item{m}{A double vector of length 1. The smallest bandwidth.}

\item{v}{A double vector of length 1. A calibration constant for weighing the variance term.}
}
\value{
A double vector of length 1.
}
\description{
Risk_hat
}
