% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandwidth_selection_GL.R,
%   R/bandwidth_selection_PCO.R
\name{est_variance}
\alias{est_variance}
\title{Estimator for the Variance Term.}
\usage{
est_variance(h, Kernel, n_obs, v)

est_variance(h, Kernel, n_obs, v)
}
\arguments{
\item{h}{A double vector of bandwidths.}

\item{Kernel}{A real function. The kernel.}

\item{n_obs}{A double vector of length 1. The number of observations.}

\item{v}{A double vector of length 1. A calibration constant.}
}
\value{
A double vector.

A double vector.
}
\description{
V_hat
source('R/bandwidth_selection.R', chdir=T)

V_hat
}
