% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandwidth_selection_PCO.R
\name{est_variance_PCO}
\alias{est_variance_PCO}
\title{Estimator for the Variance Term.}
\usage{
est_variance_PCO(h, Kernel, maxEval, n_obs)
}
\arguments{
\item{h}{A double vector of bandwidths.}

\item{Kernel}{A real function. The kernel.}

\item{maxEval}{A double vector of length 1. The maximum number of function evaluations when integrating.}

\item{n_obs}{A double vector of length 1. The number of observations.}
}
\value{
A double vector.
}
\description{
V_hat
}
