% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{get_cauchy}
\alias{get_cauchy}
\title{Get a Cauchy distribution function.}
\usage{
get_cauchy(x0 = 0, gamma = 0.5)
}
\arguments{
\item{x0}{A double vector of length 1. The location parameter.}

\item{gamma}{A double vector of length 1. The scale parameter. It should be greater than 0}
}
\value{
A vectorised function. A Cauchy distribution function.
}
\description{
Get a Cauchy distribution function.
}
