% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandwidth_selection_CV.R
\name{get_criterion_CV}
\alias{get_criterion_CV}
\title{Optimisation criterion for bandwidth selection using the Cross Validation method.}
\usage{
get_criterion_CV(Kernel, data)
}
\arguments{
\item{Kernel}{A real function. The kernel.}

\item{data}{A double vector of the sample data to use.}
}
\value{
A vectorised single-parameter function. The Cross Validation bandwidth selection
optimisation criterion.
}
\description{
Returns the Cross Validation Criterion for fixed Kernel and data only dependend on h
Minimise it to find the optimal value for h.
}
