% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandwidth_selection_GL.R
\name{get_criterion_GL}
\alias{get_criterion_GL}
\title{Optimisation criterion for bandwidth selection using the Goldenshluger-Lepski method.}
\usage{
get_criterion_GL(Kernel, data, bandwidths, c = 1, v = 1, maxEval = 1e+06)
}
\arguments{
\item{Kernel}{A real function. The kernel.}

\item{data}{A double vector of the sample data to use.}

\item{bandwidths}{A double vector containing the bandwidths to try.}

\item{c}{A double vector of length 1. A calibration constant.}

\item{v}{A double vector of length 1. A calibration constant.}

\item{maxEval}{A double vector of length 1. The maximum number of function evaluations when integrating.}
}
\value{
A vectorised single-parameter function. The Goldenshluger-Lepski bandwidth selection
optimisation criterion.
}
\description{
The Estimator for the Risk with all parameters fixed but the bandwidth h.
Minimise it to find the optimal value for h.
}
