% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{get_custom}
\alias{get_custom}
\title{Get a distribution function from a string containing a custom expression.}
\usage{
get_custom(str = "abs(x)/4 * (abs(x) <= 2)")
}
\arguments{
\item{str}{A character vector of lengh 1. A string containing a function expression using a numeric parameter named x. Must be a valid function body. The resulting function must be vectorised.}
}
\value{
A vectorised function. A custom distribution function.
}
\description{
Get a distribution function from a string containing a custom expression.
}
