% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandwidth_selection_GL.R
\name{get_double_kernel_estimator}
\alias{get_double_kernel_estimator}
\title{Double kernel estimator.}
\usage{
get_double_kernel_estimator(h, h_prime, Kernel, data)
}
\arguments{
\item{h}{A double vector of length 1. A bandwidth.}

\item{h_prime}{A double vector of length 1. Another bandwidth.}

\item{data}{A double vector of the sample data to use.}
}
\value{
A function. The double kernel estimator.
}
\description{
Double kernel estimator.
}
