% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{get_exponential}
\alias{get_exponential}
\title{Get an exponential distribution function.}
\usage{
get_exponential(lambda = 1)
}
\arguments{
\item{lambda}{A double vector of length 1. It should be bigger than 0
The rate (inverse scale) parameter.}
}
\value{
A vectorised function. An exponential distribution function.
}
\description{
Get an exponential distribution function.
}
