% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDE.R
\name{get_kde}
\alias{get_kde}
\title{Get a fix KDE.}
\usage{
get_kde(h, Kernel, data)
}
\arguments{
\item{h}{A double vector of length 1. The bandwidth.}

\item{Kernel}{A real function. The kernel.}

\item{data}{A double vector of the sample data to use.}
}
\value{
A double vector of length 1.
}
\description{
Get a fix KDE.
}
