% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{get_laplace}
\alias{get_laplace}
\title{Get a Laplace distribution function.}
\usage{
get_laplace(mu = 0, b = 1)
}
\arguments{
\item{mu}{A double vector of length 1. The mean.}

\item{b}{A double vector of length 1. The scale parameter.}
}
\value{
A vectorised function. A Laplace distribution function.
}
\description{
Get a Laplace distribution function.
}
