% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{get_uniform}
\alias{get_uniform}
\title{Get a uniform distribution function.}
\usage{
get_uniform(a = -1, b = 1)
}
\arguments{
\item{a}{A double vector of length 1. A location parameter.}

\item{b}{A double vector of length 1. A location parameter.}
}
\value{
A vectorised function. A uniform distribution function.
}
\description{
Get a uniform distribution function.
}
