% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/rejection_sample.R, R/rejection_sampling.R
=======
% Please edit documentation in R/rejection_sample.R
>>>>>>> e96bec25e8502ca9d80ad22ef322fba75463d1da
\name{helper}
\alias{helper}
\title{Constructor of a helper (distribution) object for rejection sampling.}
\usage{
helper(distribution, density)
<<<<<<< HEAD

helper(distribution, density)
=======
>>>>>>> e96bec25e8502ca9d80ad22ef322fba75463d1da
}
\arguments{
\item{distribution}{A function. A distribution function.
Must not have any non-optional parameters. Must return a double vector
of length 1 (single sample at once) when called without parameters.}

\item{density}{A function. The corresponding probability density
function to the distribution. Must accept a double vector as a
parameter. Must return a double vector of the same length.
(Vectorisation is not required.)}
}
\value{
A helper object composed of the distribution and the density functions.
<<<<<<< HEAD

A helper object composed of the distribution and the density functions.
=======
>>>>>>> e96bec25e8502ca9d80ad22ef322fba75463d1da
}
\description{
A helper distribution is a distribution function which has an
attribute 'density' which is a probability density function.
It is used for rejection sampling.
<<<<<<< HEAD

A helper distribution is a distribution function which has an
attribute 'density' which is a probability density function.
It is used for rejection sampling.
=======
>>>>>>> e96bec25e8502ca9d80ad22ef322fba75463d1da
}
