% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rejection_sample.R, R/rejection_sampling.R
\docType{data}
\name{helpers}
\alias{helpers}
\title{Helper distributions for rejection sampling.}
\format{
An object of class \code{list} of length 2.

An object of class \code{list} of length 2.
}
\usage{
helpers

helpers
}
\description{
Helper distributions for rejection sampling.

Helper distributions for rejection sampling.
}
\keyword{datasets}
