% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/rejection_sample.R, R/rejection_sampling.R
=======
% Please edit documentation in R/rejection_sample.R
>>>>>>> e96bec25e8502ca9d80ad22ef322fba75463d1da
\name{is_helper}
\alias{is_helper}
\title{Validation function for helper (distribution) objects for rejection sampling.}
\usage{
is_helper(object)
<<<<<<< HEAD

is_helper(object)
=======
>>>>>>> e96bec25e8502ca9d80ad22ef322fba75463d1da
}
\arguments{
\item{object}{An object. This object is validated to be a helper object.}
}
\value{
A logical value.
<<<<<<< HEAD

A logical value.
}
\description{
Validation function for helper (distribution) objects for rejection sampling.

Validation function for helper (distribution) objects for rejection sampling.
=======
}
\description{
Validation function for helper (distribution) objects for rejection sampling.
>>>>>>> e96bec25e8502ca9d80ad22ef322fba75463d1da
}
