% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{is_Kernel}
\alias{is_Kernel}
\title{Validation function for Kernel objects.}
\usage{
is_Kernel(object, check_values = FALSE, tolerance_rel = 1e-12)
}
\arguments{
\item{object}{Any object. This object is validated to be a Kernel.}

\item{check_values}{TRUE or FALSE. If set to TRUE, the L2norm_squared and L2norm values are
checked and the integral of the Kernel is calculated. The return value will be TRUE only if
the integral equals 1.}

\item{tolerance_rel}{A double vector of length 1. The relative tolerance for any equality checks.
This is only needed when check_values is set to TRUE.}
}
\value{
A logical.
}
\description{
Validation function for Kernel objects.
}
