% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDE-package.R
\docType{package}
\name{KDE-package}
\alias{KDE}
\alias{KDE-package}
\title{KDE: Tools for Kernel Density Estimation including Bandwidth Selection Methods}
\description{
A package containing tools for Kernel Density Estimation, including rejection sampling the to simulate data from density functions, methods for bandwidth selection, and implementation of kernels.
}
\author{
\strong{Maintainer}: Carla Sagebiel \email{carlachristin@googlemail.com}

Authors:
\itemize{
  \item Charlotte Boys
  \item Leon Patzig
}

}
\keyword{internal}
