% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{Kernel}
\alias{Kernel}
\title{Constructor for a Kernel object.}
\usage{
Kernel(func, maxEval = 1e+06)
}
\arguments{
\item{func}{A function. The kernel function.}

\item{maxEval}{A double vector of length 1. The maximum number of function evaluations when integrating.}
}
\value{
A Kernel object.
}
\description{
Constructor for a Kernel object.
}
