% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L2norm.R
\name{L2norm_squared}
\alias{L2norm_squared}
\title{Calculate L2-norm squared of a real function.}
\usage{
L2norm_squared(f)
}
\arguments{
\item{f}{A real function}
}
\value{
A double vector of length 1. The L2-norm squared of f.
}
\description{
Calculate the L2-norm squared of a real function using integrate.
}
\examples{
L2norm_squared(kernels$triangular)
}
