% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{p_sapply}
\alias{p_sapply}
\title{Use a parallelised version of sapply, if available.}
\usage{
p_sapply(X, FUN, ...)
}
\arguments{
\item{X}{A vector (atomic or list) or an `expression` object. Other objects (including classed objects) will be coerced by `base::as.list`}

\item{FUN}{A function. The function to be applied to each element of `X`. See `base::sapply` for details.}

\item{...}{Optional arguments to `FUN`.}
}
\value{
A vector.
}
\description{
Use a parallelised version of sapply, if available.
}
