% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\docType{data}
\name{pdf_factories}
\alias{pdf_factories}
\title{All probability density functions available in one list.}
\format{
An object of class \code{list} of length 6.
}
\usage{
pdf_factories
}
\description{
All probability density functions available in one list.
}
\details{
Users can use these function factories to set their own parameters.\cr
Options and defaults are as follows:  \cr
\code{pdf_factories$Cauchy(x0 = 0, gamma = 0.5)} \cr
\code{pdf_factories$Uniform(a = -1, b = 1)} \cr
\code{pdf_factories$Expo(lambda = 1)} \cr
\code{pdf_factories$Laplace(mu = 0, b = 1)} \cr
\code{pdf_factories$Custom(str = 'abs(x)/4 * (abs(x) <= 2)')}
}
\keyword{datasets}
