% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\docType{data}
\name{pdfs}
\alias{pdfs}
\title{Some probability density functions.}
\format{
An object of class \code{list} of length 4.
}
\usage{
pdfs
}
\description{
Some probability density functions.
}
\keyword{datasets}
