% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rejection_sample.R
\name{rejection_sample}
\alias{rejection_sample}
\title{Rejection sampling.}
\usage{
rejection_sample(n_obs, f, helper = helpers$normal, n_iter = 10)
}
\arguments{
\item{n_obs}{A positive integer vector of length 1. The number of observations.}

\item{f}{A probability density function to simulate observations from.}

\item{helper}{A helper (distribution) object.}

\item{n_iter}{An integer vector of length 1. A calibrating constant. The
algorithm will take an average of n_iter iterations to obtain a sample.}
}
\value{
A double vector of length n_obs. The observations.
}
\description{
Generate observations from an unknown distribution with a known density
function using the rejection sampling method.
}
