% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{sapplify}
\alias{sapplify}
\title{Vectorise a single-parameter function using sapply.}
\usage{
sapplify(f)
}
\arguments{
\item{f}{A single-parameter function.}
}
\value{
A vectorised single-parameter function.
}
\description{
Vectorise a single-parameter function using sapply.
}
