% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDE.R
\name{scaled_kernel}
\alias{scaled_kernel}
\title{Scaled kernel.}
\usage{
scaled_kernel(x, h, Kernel)
}
\arguments{
\item{x}{A double vector. The argument.}

\item{h}{A double vector of length 1. The bandwidth.}

\item{Kernel}{A real function. The kernel.}
}
\value{
A double vector.
}
\description{
Scaled kernel.
}
