% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{setup_cluster}
\alias{setup_cluster}
\title{Set up the cluster for parallel computing.}
\usage{
setup_cluster(
  n_cores = -1,
  use_parallel = NULL,
  overwrite = NULL,
  type = "FORK"
)
}
\arguments{
\item{n_cores}{A numeric vector of length 1. The number of cores to use for the cluster. When set to 0 or a negative value, a number of n_cores plus the number of cores totally available is used instead.}

\item{use_parallel}{A logical vector of length 1 or NULL. This determines whether parallalisation should be used. If set to NULL (default), parallalisation is used if the `parallel` package is installed.}

\item{overwrite}{A logical vector of length 1 or NULL. This determines whether the global cluster should be overwritten. If set to NULL (default), a warning is raised if the object is overwritten. If set to TRUE, any existing `cluster` will be overwritten silently. If set to FALSE, this method will fail throwing an error if a global `cluster` is already present.}

\item{type}{A character vector of length 1. The type of cluster to be used. See `?parallel::makeCluster` for details.}
}
\value{
An integer vector of length 1 or NULL. The actual number of cores used in the cluster. If not NULL, it will always be greater than 0.
}
\description{
Set up the cluster for parallel computing.
}
