% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandwidth_selection.R
\name{bandwidth_selection}
\alias{bandwidth_selection}
\title{Bandwidth selection using optimise.}
\usage{
bandwidth_selection(
  criterion_method,
  Kernel = kernels$gaussian,
  data,
  lower = 0.001,
  upper = 1,
  ...,
  set_up_cluster = TRUE
)
}
\arguments{
\item{criterion_method}{A string. Keyword of the bandwidth selection algorithm to use. Implemented are: 
"CV" the Cross Validation method
"GL" the Goldenshluger-Lepski method
"PCO" the Penalised Comparison to Overfitting method}

\item{Kernel}{A real function. The kernel.}

\item{data}{A double vector of the sample data to use.}

\item{lower}{A double vector of length 1. The lowest bandwidth to test.}

\item{upper}{A double vector of length 1. The greatest bandwidth to test.}

\item{...}{Optional arguments for the criterion function.}

\item{set_up_cluster}{A logical vector of length 1. Whether parallelisation should be activated if `setup_cluster` has not been run yet.}
}
\value{
A double vector of length 1. The optimal bandwidth.
}
\description{
Find the optimal value for the bandwidth h.
}
