% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandwidth_selection_PCO.R
\name{get_criterion_PCO}
\alias{get_criterion_PCO}
\title{Optimisation criterion for bandwidth selection using PCO.}
\usage{
get_criterion_PCO(Kernel, data, m, v = 1, maxEval = 1e+06)
}
\arguments{
\item{Kernel}{A real function. The kernel.}

\item{data}{A double vector of the sample data to use.}

\item{m}{A double vector of length 1. The smallest bandwidth.}

\item{v}{A double vector of length 1. A calibration constant.}

\item{maxEval}{A double vector of length 1. The maximum number of function evaluations when integrating.}
}
\value{
A vectorised single-parameter function. The PCO bandwidth selection
optimisation criterion.
}
\description{
The Estimator for the Risk with all parameters fixed but the bandwidth h.
Minimise it to find the optimal value for h.
}
