% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rejection_sample.R
\name{helper}
\alias{helper}
\title{Constructor of a helper (distribution) object for rejection sampling.}
\usage{
helper(distribution, density)
}
\arguments{
\item{distribution}{A function. A distribution function.
Must not have any non-optional parameters. Must return a double vector
of length 1 (single sample at once) when called without parameters.}

\item{density}{A function. The corresponding probability density
function to the distribution. Must accept a double vector as a
parameter. Must return a double vector of the same length.
(Vectorisation is not required.)}
}
\value{
A helper object composed of the distribution and the density functions.
}
\description{
A helper distribution is a distribution function which has an
attribute 'density' which is a probability density function.
It is used for rejection sampling.
}
