% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/rejection_sample.R, R/rejection_sampling.R
=======
% Please edit documentation in R/rejection_sample.R
>>>>>>> e96bec25e8502ca9d80ad22ef322fba75463d1da
\docType{data}
\name{helpers}
\alias{helpers}
\title{Helper distributions for rejection sampling.}
\format{
An object of class \code{list} of length 2.
<<<<<<< HEAD

An object of class \code{list} of length 2.
}
\usage{
helpers

helpers
}
\description{
Helper distributions for rejection sampling.

Helper distributions for rejection sampling.
=======
}
\usage{
helpers
}
\description{
Helper distributions for rejection sampling.
>>>>>>> e96bec25e8502ca9d80ad22ef322fba75463d1da
}
\keyword{datasets}
