% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rejection_sample.R, R/rejection_sampling.R
\name{is_helper}
\alias{is_helper}
\title{Validation function for helper (distribution) objects for rejection sampling.}
\usage{
is_helper(object)

is_helper(object)
}
\arguments{
\item{object}{An object. This object is validated to be a helper object.}
}
\value{
A logical value.

A logical value.
}
\description{
Validation function for helper (distribution) objects for rejection sampling.

Validation function for helper (distribution) objects for rejection sampling.
}
