% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{Kernel}
\alias{Kernel}
\title{Constructor for a Kernel object.}
\usage{
Kernel(func)
}
\arguments{
\item{func}{A function. The kernel function. Must be vectorised.}
}
\value{
A Kernel object.
}
\description{
Constructor for a Kernel object.
}
