% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\docType{data}
\name{pdf_factories}
\alias{pdf_factories}
\title{All probability density functions available in one list.}
\format{
An object of class \code{list} of length 6.
}
\usage{
pdf_factories
}
\description{
All probability density functions available in one list.
}
\keyword{datasets}
