% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\docType{data}
\name{pdfs}
\alias{pdfs}
\title{Some preset probability density functions.}
\format{
An object of class \code{list} of length 4.
}
\usage{
Options: \code{pdfs$Cauchy}, \code{pdfs$Uniform}, \code{pdfs$Expo}, \code{pdfs$Mix2Gauss}
}
\description{
Some preset probability density functions.
}
\keyword{datasets}
